ALL TESTS PASS
-Subclassed unwieldy Continuation4<ModuleName, Long, Long, Long> as
AbstractAvailCompiler.ProgressReporter.
-Rebuilt raw pojos to use a *descriptor* field to point to the Java
object.  No more messy weak/strong tables.  The old mechanism also
couldn't collect cycles.
-Optimized all descriptor constructors to avoid reflective lookup of
related ObjectSlots and IntegerSlots (primarily to speed up the new pogo
mechanism which creates lots of descriptor instances).
-Reworked MethodDescriptor and VariableSharedDescriptor (which implement
A_ChunkDependable) to hold a pojo containing a weak set of dependent
L2Chunks.  L2Chunks no longer have an index.
-Properties of atoms are held in a pojo-wrapped weak map.  If a property
key becomes otherwise unreachable, it is garbage collected, dropping
that (inaccessible) property.

+++ Avail/src/com/avail/descriptor/A_RawFunction.java
@@ -64,7 +64,7 @@ extends A_BasicObject
 
 	public void setStartingChunkAndReoptimizationCountdown (
 		L2Chunk chunk,
-		int countdown);
+		long countdown);
 
 	public int maxStackDepth ();
